monitoring-agent/pyproject.toml
[tool.poetry]
name = "monitoring-agent"
version = "0.1.0"
description = ""
authors = ["517ANT39 <toni.tashchilin@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psutil = "^7.2.2"
python-dotenv = "^1.2.1"
confluent-kafka = "^2.13.0"
apscheduler = "^3.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
